Move piece enumerations into single file
