--- twophase/color.py
@@ -1,10 +1,0 @@
-from enum import IntEnum
-
-
-class Color(IntEnum):
-    U = 0
-    R = 1
-    F = 2
-    D = 3
-    L = 4
-    B = 5
