--- twophase/corner.py
@@ -1,12 +1,0 @@
-from enum import IntEnum
-
-
-class Corner(IntEnum):
-    URF = 0
-    UFL = 1
-    ULB = 2
-    UBR = 3
-    DFR = 4
-    DLF = 5
-    DBL = 6
-    DRB = 7
