--- twophase/edge.py
@@ -1,16 +1,0 @@
-from enum import IntEnum
-
-
-class Edge(IntEnum):
-    UR = 0
-    UF = 1
-    UL = 2
-    UB = 3
-    DR = 4
-    DF = 5
-    DL = 6
-    DB = 7
-    FR = 8
-    FL = 9
-    BL = 10
-    BR = 11
