+++ twophase/pieces.py
@@ -1,6 +1,41 @@
 from enum import IntEnum
 
 
+class Color(IntEnum):
+    U = 0
+    R = 1
+    F = 2
+    D = 3
+    L = 4
+    B = 5
+
+
+class Corner(IntEnum):
+    URF = 0
+    UFL = 1
+    ULB = 2
+    UBR = 3
+    DFR = 4
+    DLF = 5
+    DBL = 6
+    DRB = 7
+
+
+class Edge(IntEnum):
+    UR = 0
+    UF = 1
+    UL = 2
+    UB = 3
+    DR = 4
+    DF = 5
+    DL = 6
+    DB = 7
+    FR = 8
+    FL = 9
+    BL = 10
+    BR = 11
+
+
 class Facelet(IntEnum):
     U1 = 0
     U2 = 1
